extract_text/requirements.txt
fastapi
uvicorn
pypdf
python-docx
python-pptx
openai
